Separated conversion from json string to np array into helper function. All tests passing again
